Route the OpenID discovery/JWKS/token fetches through the SSRF-hardened transport [#755]
The avatar fetch already routes every connection through a ConnectCallback that
resolves the host and connects only to a non-blocked (public) address, closing
SSRF + DNS-rebinding. The OpenID backchannel — discovery, JWKS, token, userinfo —
did not: OidcLoginService and OidcDiscoveryReader build their client from
SsoHttp.CreateClient, which returned a plain factory client with no transport
guard, so a provider endpoint (or a discovery redirect) resolving to a
private/loopback address could be reached.

Consolidate the SSRF-safe transport into the Net module and route both consumers
through it:

- SsoHttp now owns the one hardened SocketsHttpHandler (CreateHardenedHandler +
  the ConnectToAllowedAddressAsync guard, moved verbatim from AvatarService — it
  uses IpAddressClassifier, already in Net) and a named "sso-outbound" client.
- The composition root (SsoOnlyServiceRegistrator) registers that named client
  with the hardened handler, so SsoHttp.CreateClient — unchanged at its two OIDC
  call sites — now returns the hardened client in production. No change to
  OidcLoginService / OidcDiscoveryReader themselves.
- AvatarService drops its private copies of the handler + connect callback and
  builds its long-lived client on SsoHttp.CreateHardenedHandler (dedup; one SSRF
  implementation, two consumers).

The seam for tests is the named client: a stub IHttpClientFactory (the controller
harness) returns its stub for any name, and a loopback fixture returns a default
(unhardened) client for "sso-outbound", so an in-process IdP stays reachable
while production is fail-closed. New tests: CreateClient resolves the named
client, and a configuration guard pins the hardened handler's ConnectCallback /
UseProxy=false / bounded redirects so a regression that reopens the vector fails.

No behaviour change for a correctly-configured public IdP. Verified by the full
build (--warnaserror both TFMs), the suite (1595 both TFMs) and the ABI-floor
build.

Closes #755
